+++ docs/source/whatsnew.rst
@@ -5,6 +5,9 @@ What's New
 
 v4.3.11
 -------
+- BUG FIX: :meth:`RedSkyUser.user_locations` has been fixed
+- :meth:`Person.enable_call_recording()` now supports arguments specific to Webex Recording.
+- BUG FIX: :meth:`Person.disable_call_recording()` has been fixed.
 - :meth:`Recording.get_transcript()` has been added to allow the recording transcript to be obtained.
 - 4.3.11 includes my yearly clean-up so a log of new documentation, logging, and some (hopefully) small changes to the internals. I've started reworking the People class, which is the oldest one, to match with the behavior of a lot of the newer classes. At some point, this will likely mean breaking some things to replace them, but that will come in 4.4.0.
 - Added :meth:`Location.set_enhanced_emergency_calling()`